+++ Package.swift
@@ -2,8 +2,8 @@
 
 import PackageDescription
 
-let tag = "v0.1.0-rc43"
-let checksum = "f127491aff45b52416c1aeaa4b84d424d3a4fb4457f80066f5aee9b2ce474754"
+let tag = "v0.1.0-rc44"
+let checksum = "541c0e4cf0b6906209559e5f05223ffaa20fa42bebae44c694e5dcb00cdb2fa8"
 let url = "https://github.com/pubky/paykit-rs/releases/download/\(tag)/Paykit.xcframework.zip"
 
 let package = Package(